Seperate listing tables and tabs
[MAILPOET-5395]
