+++ mailpoet/assets/js/src/segments/dynamic/list.tsx
@@ -9,7 +9,7 @@ import * as ROUTES from 'segments/routes';
 import { __, _n, sprintf } from '@wordpress/i18n';
 import { TopBarWithBeamer } from 'common/top-bar/top-bar';
 import { ListingHeader } from 'segments/dynamic/list/listing-header';
-import { ListingTable } from './list/listing-table';
+import { ListingTab } from './list/listing-tab';
 
 type ColumnType = {
   name: string;
@@ -259,7 +259,7 @@ function DynamicSegmentListComponent(
     <>
       <TopBarWithBeamer hideScreenOptions />
       <ListingHeader />
-      <ListingTable />
+      <ListingTab />
       <Listing
         limit={window.mailpoet_listing_per_page}
         location={props.location}
